+++ lib/src/metric/mape.dart
@@ -3,6 +3,6 @@ part of metric;
 class _MAPEMetric implements Metric {
   const _MAPEMetric();
 
-  double getError(Vector predictedLabels, Vector origLabels) =>
-      100 / predictedLabels.length * ((origLabels - predictedLabels) / origLabels).abs(inPlace: true).sum();
+  double getError(Float32x4Vector predictedLabels, Float32x4Vector origLabels) =>
+      100 / predictedLabels.length * ((origLabels - predictedLabels) / origLabels).abs().sum();
 }