+++ lib/src/metric/rmse.dart
@@ -3,6 +3,6 @@ part of metric;
 class _RMSEMetric implements Metric {
   const _RMSEMetric();
 
-  double getError(Vector predictedLabels, Vector origLabels) =>
+  double getError(Float32x4Vector predictedLabels, Float32x4Vector origLabels) =>
     math.sqrt(((predictedLabels - origLabels).intPow(2)).mean());
 }