+++ lib/src/model_selection/validator/cross_validator.dart
@@ -1,4 +1,4 @@
-import 'package:dart_ml/src/math/vector/vector.dart';
+import 'package:dart_vector/vector.dart';
 import 'package:dart_ml/src/predictor/interface/predictor.dart';
 import 'package:dart_ml/src/metric/metric.dart';
 import 'package:dart_ml/src/data_splitter/interface/splitter.dart';
@@ -19,7 +19,7 @@ class CrossValidator {
 
   CrossValidator._internal(this._splitter);
 
-  Vector validate(Predictor predictor, List<Vector> features, Vector labels, {Metric metric}) {
+  Float32x4Vector validate(Predictor predictor, List<Float32x4Vector> features, List<double> labels, {Metric metric}) {
     if (features.length != labels.length) {
       throw new Exception('Number of features objects must be equal to the number of labels!');
     }
@@ -29,10 +29,11 @@ class CrossValidator {
     int scoreCounter = 0;
 
     for (Iterable<int> testIndices in allIndices) {
-      List<Vector> trainFeatures = new List<Vector>(features.length - testIndices.length);
-      Vector trainLabels = new Vector.zero(features.length - testIndices.length);
-      List<Vector> testFeatures = new List<Vector>(testIndices.length);
-      Vector testLabels = new Vector.zero(testIndices.length);
+      List<Float32x4Vector> trainFeatures = new List<Float32x4Vector>(features.length - testIndices.length);
+      List<double> trainLabels = new List<double>.filled(features.length - testIndices.length, 0.0);
+
+      List<Float32x4Vector> testFeatures = new List<Float32x4Vector>(testIndices.length);
+      List<double> testLabels = new List<double>.filled(testIndices.length, 0.0);
 
       int trainSamplesCounter = 0;
       int testSamplesCounter = 0;
@@ -54,6 +55,6 @@ class CrossValidator {
       scores[scoreCounter++] = predictor.test(testFeatures, testLabels, metric: metric);
     }
 
-    return new Vector.from(scores);
+    return new Float32x4Vector.from(scores);
   }
 }