+++ lib/src/predictor/interface/predictor.dart
@@ -1,9 +1,9 @@
-import 'package:dart_ml/src/math/vector/vector.dart';
+import 'package:dart_vector/vector.dart';
 import 'package:dart_ml/src/metric/metric.dart';
 
 abstract class Predictor {
   Metric get metric;
-  void train(List<Vector> features, Vector labels, {Vector weights});
-  double test(List<Vector> features, Vector origLabels, {Metric metric});
-  Vector predict(List<Vector> features);
+  void train(List<Float32x4Vector> features, List<double> labels, {Float32x4Vector weights});
+  double test(List<Float32x4Vector> features, List<double> origLabels, {Metric metric});
+  Float32x4Vector predict(List<Float32x4Vector> features);
 }