+++ lib/src/predictor/linear/classifier/gradient_classifier.dart
@@ -1,4 +1,4 @@
-import 'package:dart_ml/src/math/vector/vector.dart';
+import 'package:dart_vector/vector.dart';
 import 'package:dart_ml/src/metric/metric.dart';
 import 'package:dart_ml/src/optimizer/gradient/interface/stochastic.dart';
 import 'package:dart_ml/src/predictor/linear/base/gradient_predictor.dart';
@@ -8,6 +8,6 @@ class GradientLinearClassifier extends GradientLinearPredictor implements Classi
   GradientLinearClassifier(SGDOptimizer optimizer, Metric metric)
       : super(optimizer, metric: metric);
 
-  Vector predictProbabilities(List<Vector> features) =>
-      new Vector.from(predict(features).map((double value) => value > 0.5 ? 1.0 : 0.0));
+  Float32x4Vector predictProbabilities(List<Float32x4Vector> features) =>
+      new Float32x4Vector.from(predict(features).asList().map((double value) => value > 0.5 ? 1.0 : 0.0));
 }