+++ lib/src/score_function/linear.dart
@@ -3,5 +3,5 @@ part of score_function;
 class _LinearScore implements ScoreFunction {
   const _LinearScore();
 
-  double score(Vector w, Vector x) => w.dot(x);
+  double score(Float32x4Vector w, Float32x4Vector x) => w.dot(x);
 }